requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
databases[sqlite]>=0.7
SQLAlchemy>=2.0
passlib[bcrypt]>=1.7
python-jose>=3.3
python-multipart>=0.0.6
jinja2>=3.1
python-pptx>=0.6.21
Pillow>=9.0
pandas>=2.0
openpyxl>=3.0
requests>=2.31
